+++ rawrbox.ui/src/scripting/wrappers/button.cpp
@@ -18,7 +18,7 @@ namespace rawrbox {
 		    .addFunction("setBackgroundColor", &rawrbox::UIButton::setBackgroundColor)
 		    .addFunction("getBackgroundColor", &rawrbox::UIButton::getBackgroundColor)
 
-		    .addFunction("setText", [](rawrbox::UIButton& self, const std::string& str, std::optional<uint32_t> size) {
+		    .addFunction("setText", [](rawrbox::UIButton& self, const std::string& str, std::optional<uint16_t> size) {
 			    self.setText(str, size.value_or(16U));
 		    })
 		    .addFunction("getText", &rawrbox::UIButton::getText)
